hrm refactoring is running

+++ app/Http/Controllers/HRM/AttendanceController.php
@@ -25,6 +25,11 @@ class AttendanceController extends Controller
 
     public function index(Request $request)
     {
+        if (!auth()->user()->can('attendances_index')) {
+
+            abort(403, 'Access Forbidden.');
+        }
+
         if ($request->ajax()) {
 
             return $this->attendanceService->attendancesTable(request: $request);
@@ -39,6 +44,11 @@ class AttendanceController extends Controller
 
     public function create()
     {
+        if (!auth()->user()->can('attendances_create')) {
+
+            abort(403, 'Access Forbidden.');
+        }
+
         $departments = DB::table('hrm_departments')->get(['id', 'name']);
         $users = DB::table('users')->where('branch_id', auth()->user()->branch_id)->get(['id', 'prefix', 'name', 'last_name', 'emp_id']);
 
@@ -47,6 +57,11 @@ class AttendanceController extends Controller
 
     public function store(Request $request)
     {
+        if (!auth()->user()->can('attendances_create')) {
+
+            abort(403, 'Access Forbidden.');
+        }
+
         if ($request->user_ids == null) {
 
             return response()->json(['errorMsg' => __('Select employee first for attendance.')]);
@@ -68,6 +83,11 @@ class AttendanceController extends Controller
 
     public function edit($id)
     {
+        if (!auth()->user()->can('attendances_edit')) {
+
+            abort(403, 'Access Forbidden.');
+        }
+
         $attendance = DB::table('hrm_attendances')
             ->leftJoin('users', 'hrm_attendances.user_id', 'users.id')
             ->where('hrm_attendances.id', $id)
@@ -86,6 +106,11 @@ class AttendanceController extends Controller
 
     public function update($id, Request $request)
     {
+        if (!auth()->user()->can('attendances_edit')) {
+
+            abort(403, 'Access Forbidden.');
+        }
+
         $this->attendanceService->validation(request: $request);
         $this->attendanceService->updateAttendance(request: $request, id: $id);
         return response()->json(__('Attendances updated successfully!'));
@@ -93,6 +118,11 @@ class AttendanceController extends Controller
 
     public function delete($id, Request $request)
     {
+        if (!auth()->user()->can('attendances_delete')) {
+
+            abort(403, 'Access Forbidden.');
+        }
+
         $this->attendanceService->deleteAttendance(id: $id);
         return response()->json(__('Attendance deleted successfully'));
     }
